+++ src/modules/layerTree/tests/unit/components/LayerTree.spec.js
@@ -107,7 +107,9 @@ describe("src/modules/layerTree/components/LayerTree.vue", () => {
                 Maps: {
                     namespaced: true,
                     getters: {
-                        mode: () => mapMode
+                        mode: () => mapMode,
+                        scale: () => 500,
+                        scales: () => [500, 1000]
                     }
                 },
                 Menu: {
@@ -177,6 +179,15 @@ describe("src/modules/layerTree/components/LayerTree.vue", () => {
                 allLayerConfigsStructured: () => allLayerConfigsStructured,
                 allLayerConfigs: () => {
                     return layersBG.concat(subjectDataLayers);
+                },
+                layerConfigById: () => (id) => {
+                    if (layersBG.find(layer => layer.id === id)) {
+                        return layersBG.find(layer => layer.id === id);
+                    }
+                    if (subjectDataLayers.find(layer => layer.id === id)) {
+                        return subjectDataLayers.find(layer => layer.id === id);
+                    }
+                    return undefined;
                 },
                 portalConfig: () => {
                     return {